agentengine/pyproject.toml
[project]
name = "search-agent"
version = "0.1.0"
description = "simple search agent from 'Build an AI Agent with Gemini 3' https://youtu.be/9EGtawwvlNs"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.20.0",
]

[dependency-groups]
dev = [
    "google-adk[eval]>=1.20.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest]
pythonpath = ["."]
